test(e2e): harden flaky recurring start-date specs via shared helpers
Five recurring start-date specs flaked in CI (run 27005469443), all in the
Material datepicker start-date flow. Two root causes:

- setStartDate set the value via fill() -> press('Tab') -> toHaveValue(),
  but the input's (dateChange) handler clears innerValue to null when a
  blur-time parse loses the race with dialog bind/animation, and the one-way
  [ngModel]="innerValue()" re-renders the field empty. The prior retry
  wrapped only the fill, not the Tab-commit, so the value still vanished.
- #7423 and #7951 navigated the planner with plain page.goto('/#/planner'),
  which Angular's router occasionally drops mid-bootstrap.

Extract the helpers (duplicated 3-5x at varying robustness) into a shared
e2e/utils/recurring-task-helpers.ts: setRecurStartDate now retries the whole
fill+Tab+verify cycle via expect().toPass(), and all planner navigation goes
through the hash-drop-resistant gotoHashRoute. #6860 keeps its keyboard-typing
path but wraps it in the same retry. Also removes a waitForTimeout(800) that
violated the e2e no-fixed-wait rule.

Verified: checkFile clean; specs pass with --repeat-each=3 (18/18).
