+++ e2e/tests/recurring/recurring-start-date-epoch-bug-6860.spec.ts
@@ -111,17 +111,20 @@ test.describe('Recurring Task - Start Date Epoch Bug (#6860)', () => {
     // The app defaults to en-GB locale (DD/MM/YYYY format)
     const dateInput = repeatDialog.getByRole('textbox', { name: /start date/i });
     await expect(dateInput).toBeVisible();
-    // Click to focus the input, then select all and type the date
-    await dateInput.click();
-    await dateInput.press('Control+a');
-    await dateInput.pressSequentially('15/06/2026', { delay: 50 });
-    // Trigger change by pressing Tab to blur
-    await dateInput.press('Tab');
-
-    // 5. Verify the input retained the typed date
-    const inputValue = await dateInput.inputValue();
-    expect(inputValue).not.toBe('');
-    expect(inputValue).not.toContain('1970');
+    // The Material datepicker input intermittently drops the typed value while
+    // the dialog is still binding/animating: the blur (Tab) clears the field
+    // when the text hasn't parsed to a valid date yet, leaving the form invalid
+    // so Save stays disabled. Retry the whole type-and-commit cycle (clear +
+    // type + blur) until the input retains the typed date.
+    await expect(async () => {
+      await dateInput.fill('');
+      await dateInput.pressSequentially('15/06/2026', { delay: 50 });
+      // Trigger change by pressing Tab to blur
+      await dateInput.press('Tab');
+      const inputValue = await dateInput.inputValue();
+      expect(inputValue).not.toBe('');
+      expect(inputValue).not.toContain('1970');
+    }).toPass({ timeout: 10000 });
 
     // 6. Save
     const saveBtn = repeatDialog.getByRole('button', { name: /Save/i });
