+++ e2e/tests/recurring/repeat-edit-strands-today-bug-7951.spec.ts
@@ -1,5 +1,10 @@
-import { Locator, Page } from '@playwright/test';
 import { expect, test } from '../../fixtures/test.fixture';
+import {
+  gotoHashRoute,
+  openRecurDialog,
+  saveRecurDialog,
+  setRecurQuickSetting,
+} from '../../utils/recurring-task-helpers';
 
 /**
  * Bug: https://github.com/super-productivity/super-productivity/issues/7951
@@ -26,35 +31,6 @@ import { expect, test } from '../../fixtures/test.fixture';
 
 const FIXED_TODAY = new Date('2026-06-16T09:00:00'); // Tuesday
 
-const openRecurDialog = async (page: Page): Promise<Locator> => {
-  const recurItem = page
-    .locator('task-detail-item')
-    .filter({ has: page.locator('mat-icon', { hasText: /^repeat$/ }) });
-  await expect(recurItem).toBeVisible({ timeout: 5000 });
-  await recurItem.click();
-  const dialog = page.locator('mat-dialog-container');
-  await dialog.waitFor({ state: 'visible', timeout: 10000 });
-  return dialog;
-};
-
-const saveDialog = async (page: Page): Promise<void> => {
-  const dialog = page.locator('mat-dialog-container');
-  const saveBtn = dialog.getByRole('button', { name: /Save/i });
-  await expect(saveBtn).toBeEnabled({ timeout: 5000 });
-  await saveBtn.click();
-  await dialog.waitFor({ state: 'hidden', timeout: 10000 });
-};
-
-// Switch the "Recurring Config" quick-setting select. The option labels come
-// from en.json (Q_MONDAY_TO_FRIDAY = "Every Monday through Friday").
-const setQuickSetting = async (page: Page, optionLabel: RegExp): Promise<void> => {
-  const dialog = page.locator('mat-dialog-container');
-  await dialog.locator('mat-select').first().click();
-  const option = page.locator('mat-option').filter({ hasText: optionLabel });
-  await expect(option).toBeVisible({ timeout: 5000 });
-  await option.click();
-};
-
 test.describe('Recurring Task - Edit must not strand today (#7951)', () => {
   test('switching Daily -> Mon-Fri keeps the live instance on today', async ({
     page,
@@ -77,20 +53,19 @@ test.describe('Recurring Task - Edit must not strand today (#7951)', () => {
 
     // 2. First save: default Daily. The live instance stays on today.
     await openRecurDialog(page);
-    await saveDialog(page);
+    await saveRecurDialog(page);
 
     // The detail panel stays open after saving the repeat config. Reopen the
     // recur dialog and switch the quick setting to Mon-Fri (a schedule-affecting
     // change that still includes today, a Tuesday).
     await openRecurDialog(page);
-    await setQuickSetting(page, /Every Monday through Friday/i);
-    await saveDialog(page);
+    await setRecurQuickSetting(page, /Every Monday through Friday/i);
+    await saveRecurDialog(page);
 
     // 3. Verify in planner: the live task is on 16/6 (today), NOT 17/6 (tomorrow).
     //    The live instance renders as <planner-task>; future occurrences render
     //    as faded <planner-repeat-projection>.
-    await page.goto('/#/planner');
-    await page.waitForLoadState('networkidle');
+    await gotoHashRoute(page, '/#/planner', page.locator('planner-day').first());
 
     const dayToday = page
       .locator('planner-day')
